requirements_test.txt
Jinja2>=3.1.2
Jinja2>=3.0.3
pytest>=8.0.1
PyYAML>=5.4.1
